+++ db/migrate/20250814130010_add_faction_to_game_participations.rb
@@ -1,0 +1,9 @@
+class AddFactionToGameParticipations < ActiveRecord::Migration[8.0]
+  def up
+    add_reference :game_participations, :faction, foreign_key: { to_table: :game_factions }
+  end
+
+  def down
+    remove_reference :game_participations, :faction, foreign_key: { to_table: :game_factions }
+  end
+end