+++ db/migrate/20250814130020_add_faction_to_tournament_registrations.rb
@@ -1,0 +1,5 @@
+class AddFactionToTournamentRegistrations < ActiveRecord::Migration[8.0]
+	def change
+		add_reference :tournament_registrations, :faction, null: true, foreign_key: { to_table: :game_factions }
+	end
+end