+++ db/seeds.rb
@@ -7,3 +7,20 @@
 #   ["Action", "Comedy", "Drama", "Horror"].each do |genre_name|
 #     MovieGenre.find_or_create_by!(name: genre_name)
 #   end
+
+# Seed default factions for demo systems
+if defined?(Game::System)
+  chess = Game::System.find_by(name: 'Chess')
+  if chess
+    %w[White Black].each do |name|
+      Game::Faction.find_or_create_by!(game_system: chess, name: name)
+    end
+  end
+
+  go = Game::System.find_by(name: 'Go')
+  if go
+    %w[Black White].each do |name|
+      Game::Faction.find_or_create_by!(game_system: go, name: name)
+    end
+  end
+end