+++ test/controllers/game/events_controller_test.rb
@@ -9,12 +9,16 @@ module Game
       @system = game_systems(:chess)
       @opponent = users(:player_two)
       post session_path, params: { email_address: @user.email_address, password: 'password' }
+
+      # Ensure factions exist for system
+      @f1 = Game::Faction.find_or_create_by!(game_system: @system, name: 'White')
+      @f2 = Game::Faction.find_or_create_by!(game_system: @system, name: 'Black')
     end
 
     test 'should get new game form' do
       get new_game_event_path
       assert_response :success
-      assert_select 'h2', I18n.t('games.new.title')
+      assert_select 'h2', text: I18n.t('games.new.title')
     end
 
     test 'should create game with scores for both players' do
@@ -23,8 +27,8 @@ module Game
           event: {
             game_system_id: @system.id,
             game_participations_attributes: [
-              { user_id: @user.id, score: 21 },
-              { user_id: @opponent.id, score: 18 }
+              { user_id: @user.id, score: 21, faction_id: @f1.id },
+              { user_id: @opponent.id, score: 18, faction_id: @f2.id }
             ]
           }
         }
@@ -41,7 +45,7 @@ module Game
         }
       }
 
-      assert_response :unprocessable_entity
+      assert_response :unprocessable_content
     end
 
     test 'should not create game if a score is missing' do
@@ -49,13 +53,27 @@ module Game
         event: {
           game_system_id: @system.id,
           game_participations_attributes: [
-            { user_id: @user.id, score: 21 },
-            { user_id: @opponent.id }
+            { user_id: @user.id, score: 21, faction_id: @f1.id },
+            { user_id: @opponent.id, faction_id: @f2.id }
           ]
         }
       }
 
-      assert_response :unprocessable_entity
+      assert_response :unprocessable_content
+    end
+
+    test 'should not create game if a faction is missing' do
+      post game_events_path, params: {
+        event: {
+          game_system_id: @system.id,
+          game_participations_attributes: [
+            { user_id: @user.id, score: 21, faction_id: @f1.id },
+            { user_id: @opponent.id, score: 18 }
+          ]
+        }
+      }
+
+      assert_response :unprocessable_content
     end
 
     test 'should not create game without players' do
@@ -66,7 +84,7 @@ module Game
         }
       }
 
-      assert_response :unprocessable_entity
+      assert_response :unprocessable_content
     end
   end
 end